requirements.txt
cloudinary==1.8.0
coverage==4.4.1
da-vinci==0.2.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
django-filter==1.1.0
django-model-utils==3.0.0
djangorestframework==3.7.1
humanize==0.5.1
mock==2.0.0
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
psycopg2==2.7.3.1
python-decouple==3.1
pytz==2017.2
shortuuid==0.5.0
six==1.11.0
static3==0.7.0
urllib3==1.22
gunicorn==19.7.1
